Homebrew is an evil thing that makes life hard
Installing postgresql@12 is apparently guaranteed to fail because `libpq` gets in the way. I don't trust what removing it might do, and there's no way to tell Brew not to try linking on install, so just ignore failures of the install command. If it fails for real, the services command will fail anyway.